+++ src/App.tsx
@@ -1,8 +1,9 @@
 import { useQuery, useMutation } from "convex/react";
 import { api } from "../convex/_generated/api";
 import { Toaster } from "sonner";
-import { DraftingInterface } from "./components/DraftingInterface";
+import { DraftingInterface } from "./Components/DraftingInterface/DraftingInterface";
 import { useEffect } from "react";
+import "./App.css";
 
 export default function App() {
   const seedCharacters = useMutation(api.characters.seedCharacters);
@@ -30,16 +31,17 @@ export default function App() {
   }, [faviconUrl]);
 
   return (
-    <div className="min-h-screen bg-gray-900">
+    <div className="App min-h-screen">
       <header className="bg-gray-800 border-b border-gray-700 px-4 py-3">
         <div className="flex justify-between items-center max-w-7xl mx-auto">
           <h1 className="text-xl font-bold text-white">PvP HSR - Draft</h1>
         </div>
       </header>
       
-      <main className="max-w-7xl mx-auto p-4">
+      <main>
         <DraftingInterface />
       </main>
+      
       <Toaster />
     </div>
   );
