+++ src/components/Draft/NoImpositionLightconeSelector/NoImpositionLightconeSelector.css
@@ -1,0 +1,115 @@
+.NoImpositionLightconeSelector {
+    position: relative;
+    z-index: 20;
+}
+
+/* Container [ Search Bar + LC Cost ] */
+.NoImpositionLightconeSelector > .inputs {
+    display: flex;
+    align-items: center;
+
+    border-bottom-left-radius: 0.5rem;
+    border-bottom-right-radius: 0.5rem;
+    background-color: rgba(0, 0, 0, 0.75);
+}
+
+/* Search Bar */
+.NoImpositionLightconeSelector > .inputs > .search-bar {
+    width: 100%;
+    height: 1.875rem;
+    padding-inline: 0.125rem;
+
+    font-size: 0.813rem;
+    background: transparent;
+}
+
+/* LC Cost */
+.NoImpositionLightconeSelector > .inputs > .cost {
+    padding-inline: 0.125rem;
+
+    font-weight: bold;
+    color: var(--color-cost);
+    
+    line-height: 1.5rem; /* dictates border height */
+    border-left: 1px solid rgb(75, 85, 99);
+}
+
+/* Search Results */
+.NoImpositionLightconeSelector > .result-set {
+    position: absolute;
+    left: 0;
+    right: 0;
+    overflow-y: auto;
+    z-index: 100;
+
+    width: 120%;
+    max-height: 12rem;
+    margin-top: 0.25rem;
+
+    border-radius: 0.5rem;
+    background-color: rgba(0, 0, 0, 0.85);
+}
+
+/* Result */
+.NoImpositionLightconeSelector > .result-set > .result {
+    width: 100%;
+    text-align: start;
+
+    display: flex;
+    align-items: center;
+    gap: 0.25rem;
+
+    padding-block: 0.5rem;
+    font-size: clamp(0.875rem, 1.2vw, 1rem);
+
+    transition: all 300ms;
+}
+
+/* Result's Name */
+.NoImpositionLightconeSelector > .result-set > .result > .name {
+    flex: 1;
+    line-height: 1.25rem;
+    padding-inline: 0.25rem;
+
+    /* two lines + hides overflow */
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+    overflow: hidden;
+}
+
+/* Result's S1 Cost */
+.NoImpositionLightconeSelector > .result-set > .result > .cost {
+    flex-shrink: 0;
+    padding-inline: 0.25rem;
+
+    color: var(--color-cost);
+    font-weight: bold;
+
+    line-height: 2.25rem; /* dictates border height */
+    border-left: 1px solid rgb(75, 85, 99);
+}
+
+/* Clear LC button */
+.NoImpositionLightconeSelector > .result-set > .clear {
+    width: 100%;
+    text-align: start;
+    padding: 0.5rem;
+
+    color: gray;
+    font-size: clamp(0.875rem, 1.2vw, 1rem);
+    font-style: italic;
+
+    transition: all 300ms;
+}
+
+/* Informational result */
+.NoImpositionLightconeSelector > .result-set > .info {
+    width: 100%;
+    text-align: start;
+    padding: 0.5rem;
+
+    cursor: default;
+    color: rgb(156, 163, 175);
+    font-size: clamp(0.875rem, 1.2vw, 1rem);
+}
