+++ src/components/DraftingInterface/DraftingInterface.css
@@ -1,0 +1,45 @@
+.DraftingInterface {
+    display: flex;
+    flex-direction: column;
+    gap: 1.25rem;
+}
+
+.DraftingInterface > .tabs {
+    display: flex;
+    justify-content: center;
+    overflow: hidden; /* don't remove. */
+}
+
+.DraftingInterface > .tabs > button {
+    max-width: 10rem;
+    overflow: hidden;
+
+    flex: 1;
+    padding: 0.75rem 0.125rem;
+
+    font-size: clamp(1rem, 2.1vw, 1.25rem);
+    font-weight: bold;
+    letter-spacing: -0.75px;
+
+    color: rgb(209, 213, 219);
+    background-color: rgb(55, 65, 81);
+
+    transition: all 300ms;
+}
+.DraftingInterface > .tabs > button:hover {
+    background-color: rgb(75, 85, 99);
+}
+
+.DraftingInterface > .main {
+    display: grid;
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+    gap: 1.25rem;
+}
+
+/* Apply on mobile+tablet screens */
+@media screen and (max-width: 1040px) {
+    /* Single grid column in smaller screens */
+    .DraftingInterface > .main {
+        grid-template-columns: repeat(1, minmax(0, 1fr));
+    }
+}
