+++ src/components/LightconeSelector/LightconeSelector.css
@@ -1,0 +1,103 @@
+.LightconeSelector {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 0.5rem;
+}
+
+/* Lightcone */
+.LightconeSelector > .lightcone {
+    flex: 1 80%;
+    height: 3rem;
+
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 0.25rem;
+
+    padding: 0.5rem;
+}
+
+/* Lightcone's name */
+.LightconeSelector > .lightcone > .lc-name {
+    font-size: clamp(0.875rem, 1.8vw, 1rem);
+    line-height: 1.125rem;
+
+    /* two lines + hides overflow */
+    max-width: calc(100% - 0.5rem);
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+    overflow: hidden;
+}
+
+/* LC options (cost + clear button) */
+.LightconeSelector > .lightcone > .lc-info {
+    display: flex;
+    gap: 0.5rem;
+}
+.LightconeSelector > .lightcone > .lc-info > .cost {
+    color: var(--color-cost);
+    font-size: clamp(1rem, 2.1vw, 1.25rem);
+}
+
+/* Lightcone's imposition */
+.LightconeSelector > .imposition {
+    flex: 1 20%;
+
+    padding: 0.375rem;
+
+    cursor: pointer;
+    font-size: clamp(0.75rem, 2.1vw, 0.875rem);
+    font-weight: bold;
+}
+
+.LightconeSelector > .lightcone-search {
+    position: relative;
+    width: 100%;
+    height: 3rem;
+}
+
+.LightconeSelector > .lightcone-search > .lc-input {
+    width: 100%;
+    height: 100%;
+
+    font-size: clamp(0.875rem, 1.8vw, 1rem);
+}
+.LightconeSelector > .lightcone-search > .result-set {
+    padding-inline: 0.25rem;
+}
+.LightconeSelector > .lightcone-search > .result-set > .result {
+    display: flex;
+    align-items: center;
+
+    padding: 0.5rem 0.25rem;
+
+    font-size: clamp(0.875rem, 2.1vw, 1rem);
+}
+.LightconeSelector > .lightcone-search > .result-set > .result > .lc-name {
+    /* two lines + hides overflow */
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+    overflow: hidden;
+}
+.LightconeSelector > .lightcone-search > .result-set > .result > .cost {
+    font-weight: bold;
+}
+
+/* Apply on mobile+tablet screens */
+@media screen and (max-width: 1040px) {
+    .LightconeSelector {
+        flex-direction: column;
+    }
+    .LightconeSelector > .imposition {
+        /* 
+        will move imposition select to character's top right 
+
+        position: absolute;
+        right: 0.325rem;
+        top: 4.875rem;
+        z-index: 20; */
+    }
+}
